+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Ability to change default upload folder using `/settings` command.
 - "Report a problem" button for internal server error message.
 - Automatic redirect from authorization page to the bot.